refactor: new packages into src/
