+++ src/config/config_test.go
@@ -1,10 +1,12 @@
-package main
+package config_test
 
 import (
 	"fmt"
 	"os"
 	"testing"
 
+	. "github.com/fljdin/dispatch/src/config"
+	. "github.com/fljdin/dispatch/src/models"
 	"github.com/stretchr/testify/assert"
 )
