+++ src/models/taskresult.go
@@ -1,0 +1,13 @@
+package models
+
+import "time"
+
+type TaskResult struct {
+	ID        int
+	StartTime time.Time
+	EndTime   time.Time
+	Elapsed   time.Duration
+	Status    int
+	Output    string
+	Error     string
+}
