+++ src/parser/parser_test.go
@@ -1,9 +1,10 @@
-package main
+package parser_test
 
 import (
 	"os"
 	"testing"
 
+	. "github.com/fljdin/dispatch/src/parser"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -88,13 +89,4 @@ func TestParserFromSqlFile(t *testing.T) {
 	queries := parser.Parse()
 
 	assert.Equal(t, 1, len(queries))
-}
-
-func TestParserFromInvalidParseType(t *testing.T) {
-	_, err := NewParserBuilder("unknown").
-		Build()
-
-	if assert.NotEqual(t, nil, err) {
-		assert.Contains(t, err.Error(), "invalid type for parsing file")
-	}
 }
