+++ src/parser/parserbuilder.go
@@ -1,0 +1,42 @@
+package parser
+
+import (
+	"fmt"
+	"os"
+)
+
+type ParserBuilder struct {
+	parser Parser
+	err    error
+}
+
+func NewParserBuilder(pt string) *ParserBuilder {
+	return &ParserBuilder{
+		parser: Parser{
+			Type: pt,
+		},
+	}
+}
+
+func (pb *ParserBuilder) WithContent(content string) *ParserBuilder {
+	pb.parser.Content = content
+	return pb
+}
+
+func (pb *ParserBuilder) FromFile(filename string) *ParserBuilder {
+	data, err := os.ReadFile(filename)
+	if err != nil {
+		pb.err = fmt.Errorf("error reading file: %w", err)
+	}
+
+	pb.parser.Content = pb.WithContent(string(data)).parser.Content
+	return pb
+}
+
+func (pb *ParserBuilder) Build() (Parser, error) {
+	if err := pb.parser.VerifyType(); err != nil {
+		pb.err = err
+	}
+
+	return pb.parser, pb.err
+}
